+++ cpp/src/CMakeLists.txt
@@ -91,7 +91,7 @@ if(WASM)
         $<TARGET_OBJECTS:stdlib_aes128_objects>
         $<TARGET_OBJECTS:stdlib_merkle_tree_objects>
         $<TARGET_OBJECTS:acir_format_objects>
-        $<TARGET_OBJECTS:proofs_objects>
+        $<TARGET_OBJECTS:acir_proofs_objects>
     )
 
     # With binaryen installed, it seems its wasm backend optimiser gets invoked automatically.
@@ -151,7 +151,7 @@ if(WASM)
         $<TARGET_OBJECTS:stdlib_aes128_objects>
         $<TARGET_OBJECTS:stdlib_merkle_tree_objects>
         $<TARGET_OBJECTS:acir_format_objects>
-        $<TARGET_OBJECTS:proofs_objects>
+        $<TARGET_OBJECTS:acir_proofs_objects>
     )
 else()
     # For use when compiling dependent cpp projects
@@ -183,7 +183,7 @@ else()
         $<TARGET_OBJECTS:stdlib_aes128_objects>
         $<TARGET_OBJECTS:stdlib_merkle_tree_objects>
         $<TARGET_OBJECTS:acir_format_objects>
-        $<TARGET_OBJECTS:proofs_objects>
+        $<TARGET_OBJECTS:acir_proofs_objects>
         $<TARGET_OBJECTS:env_objects>
     )
