+++ cpp/src/barretenberg/dsl/acir_proofs/acir_proofs.hpp
@@ -1,7 +1,7 @@
 #include <cstdint>
 #include <cstddef>
 
-namespace proofs {
+namespace acir_proofs {
 
 uint32_t get_exact_circuit_size(uint8_t const* constraint_system_buf);
 uint32_t get_total_circuit_size(uint8_t const* constraint_system_buf);
@@ -16,4 +16,4 @@ size_t new_proof(void* pippenger,
 bool verify_proof(
     uint8_t const* g2x, uint8_t const* vk_buf, uint8_t const* constraint_system_buf, uint8_t* proof, uint32_t length);
 
-} // namespace proofs
+} // namespace acir_proofs
